FPAD-8110: Move InterventionName
This helps with bundling the sdk
