+++ src/data-types/constants.ts
@@ -140,10 +140,3 @@ export enum InterventionState {
   MITIGATED = 'MITIGATED',
   REMOVED = 'REMOVED',
 }
-
-export enum InterventionName {
-  PERMANENT_SUSPENSION = 'PERMANENT_SUSPENSION',
-  TEMPORARY_SUSPENSION = 'TEMPORARY_SUSPENSION',
-  RESET_PASSWORD = 'RESET_PASSWORD', //pragma: allowlist secret
-  REPROVE_IDENTITY = 'REPROVE_IDENTITY',
-}
