+++ src/data-types/intervention-name.ts
@@ -1,0 +1,6 @@
+export enum InterventionName {
+  PERMANENT_SUSPENSION = 'PERMANENT_SUSPENSION',
+  TEMPORARY_SUSPENSION = 'TEMPORARY_SUSPENSION',
+  RESET_PASSWORD = 'RESET_PASSWORD', //pragma: allowlist secret
+  REPROVE_IDENTITY = 'REPROVE_IDENTITY',
+}
